+++ packages/transport/src/utils/bridgeApiResult.ts
@@ -24,9 +24,8 @@ export function info(res: UnknownPayload) {
     if (typeof version !== 'string') {
         return error({ code: ERRORS.WRONG_RESULT_TYPE });
     }
-    const protocolMessages = !!res.protocolMessages;
 
-    return success({ version, protocolMessages });
+    return success({ version });
 }
 
 export function devices(res: UnknownPayload) {